Relocate sig reqs pre-handlers (#4559)
