+++ hedera-node/hedera-app-spi/src/main/java/com/hedera/node/app/spi/PreHandleDispatcher.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022 Hedera Hashgraph, LLC
+ * Copyright (C) 2022-2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,6 +24,7 @@ import edu.umd.cs.findbugs.annotations.NonNull;
  * A {@code PreHandleDispatcher} takes a validated transaction and dispatches it to the correct
  * handler
  */
+@FunctionalInterface
 public interface PreHandleDispatcher {
     /**
      * Dispatch a request. It is forwarded to the correct handler, which takes care of the specific
