+++ hedera-node/hedera-token-service-impl/src/main/java/com/hedera/node/app/service/token/impl/handlers/CryptoDeleteHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022 Hedera Hashgraph, LLC
+ * Copyright (C) 2022-2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,10 @@
  */
 package com.hedera.node.app.service.token.impl.handlers;
 
+import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_TRANSFER_ACCOUNT_ID;
+
 import com.hedera.node.app.service.token.impl.ReadableAccountStore;
+import com.hedera.node.app.spi.meta.SigTransactionMetadataBuilder;
 import com.hedera.node.app.spi.meta.TransactionMetadata;
 import com.hedera.node.app.spi.workflows.TransactionHandler;
 import com.hederahashgraph.api.proto.java.AccountID;
@@ -27,18 +30,13 @@ import edu.umd.cs.findbugs.annotations.NonNull;
  * com.hederahashgraph.api.proto.java.HederaFunctionality#CryptoDelete}.
  */
 public class CryptoDeleteHandler implements TransactionHandler {
-
     /**
-     * This method is called during the pre-handle workflow.
+     * Pre-handles a {@link
+     * com.hederahashgraph.api.proto.java.HederaFunctionality#CryptoDeleteAllowance} transaction,
+     * returning the metadata required to, at minimum, validate the signatures of all required
+     * signing keys.
      *
-     * <p>Typically, this method validates the {@link TransactionBody} semantically, gathers all
-     * required keys, warms the cache, and creates the {@link TransactionMetadata} that is used in
-     * the handle stage.
-     *
-     * <p>Please note: the method signature is just a placeholder which is most likely going to
-     * change.
-     *
-     * @param txBody the {@link TransactionBody} with the transaction data
+     * @param txn the {@link TransactionBody} with the transaction data
      * @param payer the {@link AccountID} of the payer
      * @param accountStore the {@link ReadableAccountStore} with the current data
      * @return the {@link TransactionMetadata} with all information that needs to be passed to
@@ -46,10 +44,20 @@ public class CryptoDeleteHandler implements TransactionHandler {
      * @throws NullPointerException if one of the arguments is {@code null}
      */
     public TransactionMetadata preHandle(
-            @NonNull final TransactionBody txBody,
+            @NonNull final TransactionBody txn,
             @NonNull final AccountID payer,
             @NonNull final ReadableAccountStore accountStore) {
-        throw new UnsupportedOperationException("Not implemented");
+        final var op = txn.getCryptoDelete();
+        final var deleteAccountId = op.getDeleteAccountID();
+        final var transferAccountId = op.getTransferAccountID();
+        final var meta =
+                new SigTransactionMetadataBuilder(accountStore)
+                        .payerKeyFor(payer)
+                        .txnBody(txn)
+                        .addNonPayerKey(deleteAccountId)
+                        .addNonPayerKeyIfReceiverSigRequired(
+                                transferAccountId, INVALID_TRANSFER_ACCOUNT_ID);
+        return meta.build();
     }
 
     /**
