+++ hedera-node/hedera-token-service/src/main/java/com/hedera/node/app/service/token/CryptoService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022 Hedera Hashgraph, LLC
+ * Copyright (C) 2022-2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -25,9 +25,10 @@ import java.util.ServiceLoader;
  * transactions and queries defined in the "CryptoService" protobuf service. The {@code
  * CryptoService} is used extensively by the core application workflows to implement transaction
  * handling, since all transactions and most queries involve payments and thus the transfer of HBAR
- * from one account to another. A {@link CryptoPreTransactionHandler} contains API for all
- * transactions related to crypto (and token) transfers, as well as some additional API needed by
- * the core application to apply payments and compute rewards.
+ * from one account to another. The {@link com.hedera.node.app.spi.workflows.TransactionHandler}
+ * implementations for the crypto service provide the API for all transactions related to crypto
+ * (and token) transfers, as well as some additional API needed by the core application to apply
+ * payments and compute rewards.
  *
  * <p>Implements the HAPI <a
  * href="https://github.com/hashgraph/hedera-protobufs/blob/main/services/crypto_service.proto">Crypto
